requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
# black @ file:///C:/ci/black_1660221726201/work
# certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.1.0
# click @ file:///C:/ci/click_1646056762388/work
# colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
customtkinter==5.1.2
darkdetect==0.8.0
defusedxml==0.7.1
fonttools==4.39.2
fpdf2==2.6.1
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
mypy-extensions==0.4.3
openai==0.27.2
# pathspec @ file:///C:/b/abs_9cu5_2yb3i/croot/pathspec_1674681579249/work
pdfkit==1.0.0
pefile==2023.2.7
Pillow==9.4.0
# platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
python-docx==0.8.11
pywin32-ctypes==0.2.0
requests==2.28.2
# tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
tqdm==4.65.0
urllib3==1.26.15
wincertstore==0.2
wkhtmltopdf==0.2
yarl==1.8.2
